# src/api/handlers/risk_handlers.go
package handlers

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/Ceesaxp/autonomous-content-service/src/domain/entities"
	"github.com/Ceesaxp/autonomous-content-service/src/services/risk_management"
	"github.com/google/uuid"
	"github.com/gorilla/mux"
)

// RiskHandlers contains handlers for risk management endpoints
type RiskHandlers struct {
	riskService risk_management.RiskManagementService
}

// NewRiskHandlers creates a new risk handlers instance
func NewRiskHandlers(riskService risk_management.RiskManagementService) *RiskHandlers {
	return &RiskHandlers{
		riskService: riskService,
	}
}

// RegisterRoutes registers all risk management routes
func (h *RiskHandlers) RegisterRoutes(router *mux.Router) {
	// Risk endpoints
	router.HandleFunc("/api/v1/risks", h.listRisks).Methods("GET")
	router.HandleFunc("/api/v1/risks", h.createRisk).Methods("POST")
	router.HandleFunc("/api/v1/risks/{id}", h.getRisk).Methods("GET")
	router.HandleFunc("/api/v1/risks/{id}", h.updateRisk).Methods("PUT")
	router.HandleFunc("/api/v1/risks/{id}", h.deleteRisk).Methods("DELETE")
	router.HandleFunc("/api/v1/risks/{id}/assess", h.assessRisk).Methods("POST")
	router.HandleFunc("/api/v1/risks/{id}/mitigate", h.mitigateRisk).Methods("POST")

	// Incident endpoints
	router.HandleFunc("/api/v1/incidents", h.listIncidents).Methods("GET")
	router.HandleFunc("/api/v1/incidents", h.createIncident).Methods("POST")
	router.HandleFunc("/api/v1/incidents/{id}", h.getIncident).Methods("GET")
	router.HandleFunc("/api/v1/incidents/{id}", h.updateIncident).Methods("PUT")
	router.HandleFunc("/api/v1/incidents/{id}/respond", h.respondToIncident).Methods("POST")

	// Vulnerability endpoints
	router.HandleFunc("/api/v1/vulnerabilities", h.listVulnerabilities).Methods("GET")
	router.HandleFunc("/api/v1/vulnerabilities/scan", h.runVulnerabilityScan).Methods("POST")
	router.HandleFunc("/api/v1/vulnerabilities/{id}", h.getVulnerability).Methods("GET")
	router.HandleFunc("/api/v1/vulnerabilities/{id}/fix", h.fixVulnerability).Methods("POST")

	// Backup endpoints
	router.HandleFunc("/api/v1/backups", h.listBackups).Methods("GET")
	router.HandleFunc("/api/v1/backups", h.createBackup).Methods("POST")
	router.HandleFunc("/api/v1/backups/{id}", h.getBackup).Methods("GET")
	router.HandleFunc("/api/v1/backups/{id}/restore", h.restoreBackup).Methods("POST")
	router.HandleFunc("/api/v1/backups/{id}/verify", h.verifyBackup).Methods("POST")

	// System monitoring endpoints
	router.HandleFunc("/api/v1/system/health", h.getSystemHealth).Methods("GET")
	router.HandleFunc("/api/v1/system/risks", h.getSystemRisks).Methods("GET")
	router.HandleFunc("/api/v1/system/dependencies", h.getDependencies).Methods("GET")
	router.HandleFunc("/api/v1/system/security-scan", h.runSecurityScan).Methods("POST")
}

// Risk Management Handlers

func (h *RiskHandlers) listRisks(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	
	// Parse query parameters
	category := r.URL.Query().Get("category")
	severity := r.URL.Query().Get("severity")
	status := r.URL.Query().Get("status")
	offset, _ := strconv.Atoi(r.URL.Query().Get("offset"))
	limit, _ := strconv.Atoi(r.URL.Query().Get("limit"))
	if limit == 0 {
		limit = 20
	}

	var risks []*entities.Risk
	var total int
	var err error

	// Filter by category, severity, or status if provided
	if category != "" {
		risks, total, err = h.riskService.GetRisksByCategory(ctx, entities.RiskCategory(category), offset, limit)
	} else if severity != "" {
		risks, total, err = h.riskService.GetRisksBySeverity(ctx, entities.RiskSeverity(severity), offset, limit)
	} else if status != "" {
		risks, total, err = h.riskService.GetRisksByStatus(ctx, entities.RiskStatus(status), offset, limit)
	} else {
		// Get all risks - need to implement this method
		http.Error(w, "General list endpoint not implemented", http.StatusNotImplemented)
		return
	}

	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	response := map[string]interface{}{
		"risks":  risks,
		"total":  total,
		"offset": offset,
		"limit":  limit,
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(response); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) getRisk(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid risk ID", http.StatusBadRequest)
		return
	}

	risk, err := h.riskService.GetRiskByID(r.Context(), id)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	if risk == nil {
		http.Error(w, "Risk not found", http.StatusNotFound)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(risk); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) createRisk(w http.ResponseWriter, r *http.Request) {
	var risk entities.Risk
	if err := json.NewDecoder(r.Body).Decode(&risk); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if err := h.riskService.CreateRisk(r.Context(), &risk); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	if err := json.NewEncoder(w).Encode(risk); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) updateRisk(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid risk ID", http.StatusBadRequest)
		return
	}

	var risk entities.Risk
	if err := json.NewDecoder(r.Body).Decode(&risk); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	risk.ID = id
	if err := h.riskService.UpdateRisk(r.Context(), &risk); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(risk); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) deleteRisk(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid risk ID", http.StatusBadRequest)
		return
	}

	if err := h.riskService.DeleteRisk(r.Context(), id); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

func (h *RiskHandlers) assessRisk(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid risk ID", http.StatusBadRequest)
		return
	}

	assessment, err := h.riskService.AssessRisk(r.Context(), id)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(assessment); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) mitigateRisk(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid risk ID", http.StatusBadRequest)
		return
	}

	var request struct {
		Actions []string `json:"actions"`
	}
	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if err := h.riskService.MitigateRisk(r.Context(), id, request.Actions); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(map[string]string{"status": "mitigation_initiated"}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

// Incident Management Handlers

func (h *RiskHandlers) listIncidents(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to listRisks
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) getIncident(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to getRisk
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) createIncident(w http.ResponseWriter, r *http.Request) {
	var incident entities.Incident
	if err := json.NewDecoder(r.Body).Decode(&incident); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if err := h.riskService.CreateIncident(r.Context(), &incident); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	if err := json.NewEncoder(w).Encode(incident); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) updateIncident(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to updateRisk
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) respondToIncident(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid incident ID", http.StatusBadRequest)
		return
	}

	if err := h.riskService.RespondToIncident(r.Context(), id); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(map[string]string{"status": "response_initiated"}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

// Vulnerability Management Handlers

func (h *RiskHandlers) listVulnerabilities(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to listRisks
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) getVulnerability(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to getRisk
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) runVulnerabilityScan(w http.ResponseWriter, r *http.Request) {
	vulnerabilities, err := h.riskService.ScanVulnerabilities(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(map[string]interface{}{
		"vulnerabilities": vulnerabilities,
		"scan_completed": true,
	}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) fixVulnerability(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid vulnerability ID", http.StatusBadRequest)
		return
	}

	if err := h.riskService.FixVulnerability(r.Context(), id); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(map[string]string{"status": "fix_applied"}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

// Backup Management Handlers

func (h *RiskHandlers) listBackups(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to listRisks
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) getBackup(w http.ResponseWriter, r *http.Request) {
	// Similar implementation to getRisk
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}

func (h *RiskHandlers) createBackup(w http.ResponseWriter, r *http.Request) {
	var request struct {
		Name string `json:"name"`
		Type string `json:"type"`
	}
	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	backup, err := h.riskService.CreateBackup(r.Context(), request.Name, request.Type)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	if err := json.NewEncoder(w).Encode(backup); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) restoreBackup(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid backup ID", http.StatusBadRequest)
		return
	}

	if err := h.riskService.RestoreBackup(r.Context(), id); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(map[string]string{"status": "restore_initiated"}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) verifyBackup(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := uuid.Parse(vars["id"])
	if err != nil {
		http.Error(w, "Invalid backup ID", http.StatusBadRequest)
		return
	}

	isValid, err := h.riskService.VerifyBackup(r.Context(), id)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(map[string]interface{}{
		"valid": isValid,
		"verified": true,
	}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

// System Monitoring Handlers

func (h *RiskHandlers) getSystemHealth(w http.ResponseWriter, r *http.Request) {
	health, err := h.riskService.GetSystemHealth(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(health); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) getSystemRisks(w http.ResponseWriter, r *http.Request) {
	risks, err := h.riskService.GetSystemRisks(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(map[string]interface{}{
		"risks": risks,
		"total": len(risks),
	}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) getDependencies(w http.ResponseWriter, r *http.Request) {
	dependencies, err := h.riskService.GetDependencies(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(map[string]interface{}{
		"dependencies": dependencies,
		"total": len(dependencies),
	}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}

func (h *RiskHandlers) runSecurityScan(w http.ResponseWriter, r *http.Request) {
	results, err := h.riskService.RunSecurityScan(r.Context())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(map[string]interface{}{
		"scan_results": results,
		"scan_completed": true,
	}); err != nil {
		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		return
	}
}