# src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Autonomous Content Creation Service API
  description: |
    API for the Autonomous Content Creation Service, providing endpoints for client management,
    project management, content creation, and financial operations.
  version: 1.0.0
  contact:
    name: Autonomous Content Service
    url: https://github.com/Ceesaxp/autonomous-content-service
servers:
  - url: /api/v1
    description: API server
security:
  - bearerAuth: []
...
paths:
...
  /risks:
    get:
      summary: List risks
      description: Retrieve a paginated list of risks with optional filtering
      tags:
        - Risk Management
      parameters:
        - name: category
          in: query
          description: Filter by risk category
          schema:
            type: string
            enum: [Content, Financial, Operational, Security, Legal, Reputation]
        - name: severity
          in: query
          description: Filter by risk severity
          schema:
            type: string
            enum: [Critical, High, Medium, Low]
        - name: status
          in: query
          description: Filter by risk status
          schema:
            type: string
            enum: [Identified, Assessing, Mitigating, Monitoring, Resolved, Accepted]
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of risks
          content:
            application/json:
              schema:
                type: object
                properties:
                  risks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Risk'
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
    post:
      summary: Create a new risk
      description: Create a new risk entry in the system
      tags:
        - Risk Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Risk'
      responses:
        '201':
          description: Risk created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
  /risks/{id}:
    get:
      summary: Get risk by ID
      description: Retrieve a specific risk by its ID
      tags:
        - Risk Management
      parameters:
        - name: id
          in: path
          required: true
          description: Risk ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Risk details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '404':
          description: Risk not found
  /risks/{id}/assess:
    post:
      summary: Assess risk
      description: Perform risk assessment and update risk scoring
      tags:
        - Risk Management
      parameters:
        - name: id
          in: path
          required: true
          description: Risk ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Risk assessment completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Risk:
      type: object
      required:
        - id
        - title
        - description
        - category
        - severity
        - likelihood
        - impact
        - status
        - identifiedAt
        - lastAssessment
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [Content, Financial, Operational, Security, Legal, Reputation]
        severity:
          type: string
          enum: [Critical, High, Medium, Low]
        likelihood:
          type: number
          minimum: 0
          maximum: 1
          description: Probability of risk occurrence (0.0 to 1.0)
        impact:
          type: number
          minimum: 0
          maximum: 1
          description: Impact magnitude if risk occurs (0.0 to 1.0)
        status:
          type: string
          enum: [Identified, Assessing, Mitigating, Monitoring, Resolved, Accepted]
        metadata:
          type: object
          description: Additional risk-specific metadata
        mitigationActions:
          type: array
          items:
            type: string
        ownerId:
          type: string
          format: uuid
        identifiedAt:
          type: string
          format: date-time
        lastAssessment:
          type: string
          format: date-time
        resolutionDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...